Add a function that hooks into `upgrader_process_complete`.
When plugin or theme auto-updates are enabled on the General > Settings page (enabled is also the default setting), a newly installed plugin or theme will automatically be added to the new WordPress Core auto-update option. This allows site owners to individually disable auto-updates for certain plugins or themes if desired using the new Core interface.